+++ spec/hatchet/profile_d_scripts_spec.rb
@@ -19,8 +19,8 @@ RSpec.describe '.profile.d/ scripts' do
           GUNICORN_CMD_ARGS=--access-logfile -
           HOME=/app
           LANG=en_US.UTF-8
-          LD_LIBRARY_PATH=/app/.heroku/vendor/lib:/app/.heroku/python/lib:
-          LIBRARY_PATH=/app/.heroku/vendor/lib:/app/.heroku/python/lib:
+          LD_LIBRARY_PATH=/app/.heroku/python/lib:
+          LIBRARY_PATH=/app/.heroku/python/lib:
           PATH=/app/.heroku/python/bin:/usr/local/bin:/usr/bin:/bin
           PWD=/app
           PYTHONHASHSEED=random
@@ -55,8 +55,8 @@ RSpec.describe '.profile.d/ scripts' do
           GUNICORN_CMD_ARGS=this-should-be-preserved
           HOME=/app
           LANG=C.UTF-8
-          LD_LIBRARY_PATH=/app/.heroku/vendor/lib:/app/.heroku/python/lib:/this-should-be-preserved
-          LIBRARY_PATH=/app/.heroku/vendor/lib:/app/.heroku/python/lib:/this-should-be-preserved
+          LD_LIBRARY_PATH=/app/.heroku/python/lib:/this-should-be-preserved
+          LIBRARY_PATH=/app/.heroku/python/lib:/this-should-be-preserved
           PATH=/app/.heroku/python/bin:/this-should-be-preserved:/usr/local/bin:/usr/bin:/bin
           PWD=/app
           PYTHONHASHSEED=this-should-be-preserved